+++ index.html
@@ -10,6 +10,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link rel="stylesheet" href="./css/style.css">
+    <link rel="icon" href="./src/logo2.png">
     <title>Tienda Online - TomaMate</title>
 </head>
 
@@ -58,7 +59,7 @@
             <a href="./subpaginas/bombillas/bombillas.html">
                 <div class="col">
                     <div class="card-producto">
-                        <img src="./src/bombilla.png" class="img-fluid" alt="Bombillas">
+                        <img src="./src/bombillas/bombilla.png" class="img-fluid" alt="Bombillas">
                         <div class="titulo-producto">Bombillas</div>
                     </div>
                 </div>
@@ -66,7 +67,7 @@
             <a href="./subpaginas/grabados/grabados.html">
                 <div class="col">
                     <div class="card-producto">
-                        <img src="./src/Grabado.png" class="img-fluid" alt="Grabados">
+                        <img src="./src/grabados/Grabado.png" class="img-fluid" alt="Grabados">
                         <div class="titulo-producto">Grabados</div>
                     </div>
                 </div>
@@ -74,7 +75,7 @@
             <a href="./subpaginas/yerba/yerba.html">
                 <div class="col">
                     <div class="card-producto">
-                        <img src="./src/yerba.png" class="img-fluid" alt="Yerba">
+                        <img src="./src/yerbas/yerba.png" class="img-fluid" alt="Yerba">
                         <div class="titulo-producto">Yerba</div>
                     </div>
                 </div>
@@ -82,7 +83,7 @@
             <a href="./subpaginas/materas/materas.html">
                 <div class="col">
                     <div class="card-producto">
-                        <img src="./src/matera.png" class="img-fluid" alt="Materas">
+                        <img src="./src/materas/matera.png" class="img-fluid" alt="Materas">
                         <div class="titulo-producto">Materas</div>
                     </div>
                 </div>
@@ -90,7 +91,7 @@
             <a href="./subpaginas/termos/termos.html">
                 <div class="col">
                     <div class="card-producto">
-                        <img src="./src/termo.png" class="img-fluid" alt="Termos">
+                        <img src="./src/termos/termo.png" class="img-fluid" alt="Termos">
                         <div class="titulo-producto">Termos</div>
                     </div>
                 </div>
