+++ src/bombillas/bombilla.png
