+++ subpaginas/grabados/grabados.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
     <link rel="stylesheet" href="../../css/style.css">
-
+    <link rel="icon" href="../../src/logo2.png">
     <title>Tienda Online - TomaMate</title>
 </head>
 
@@ -43,19 +43,11 @@
         </div>
     </header>
 
-    <main class="container my-5">
-        <div class="row row-cols-1 row-cols-md-3 g-4">
-            <div class="col">
-                <div class="card" style="width: 18rem;">
-                    <img src="../../src/Grabado.png" class="card-img-top" alt="producto">
-                    <div class="card-body">
-                        <h5 class="card-title">Mate grabado personalizado</h5>
-                        <p class="card-text">$1.999</p>
-                        <button type="button" class="btn btn-primary btn-compra">Agregar al carrito</button>
-                    </div>
-                </div>
-            </div>
-    </main>
+    <!-- Contenedor de productos -->
+    <div class="container my-5">
+        <div id="productos-container" class="row g-4">
+        </div>
+    </div>
 
 
      <footer>
@@ -83,6 +75,8 @@
 
     </footer>
     <script type="module" src="../../carrito.js"></script>
+    <script type="module" src="../productos.js"></script>
+    <script type="module" src="grabados_app.js"></script>
 </body>
 
 </html>
