+++ subpaginas/mates/mates.html
@@ -9,8 +9,9 @@
     <link href="https://fonts.googleapis.com/css2?family=Corinthia:wght@400;700&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="icon" href="./src/logo2.png">
     <link rel="stylesheet" href="../../css/style.css">
-
+    <link rel="icon" href="../../src/logo2.png">
     <title>Tienda Online - TomaMate</title>
 </head>
 
@@ -76,6 +77,7 @@
 
     </footer>
     <script src="../../carrito.js"></script>
+    <script type="module" src="../productos.js"></script>
     <script type="module" src="mates_app.js"></script>
 </body>
