+++ subpaginas/mates/mates_app.js
@@ -1,4 +1,4 @@
-import { agregarAlCarrito } from '../../carrito.js';
+import { mostrar_productos } from '../productos.js';
 
 const productos = [
   { id: 1, nombre: "Mate de Madera", precio: 1999, imagen: "../../src/mates/mate.png" },
@@ -7,31 +7,4 @@ const productos = [
   { id: 4, nombre: "Mate Enlozado chico", precio: 1500, imagen: "../../src/mates/enlozado.png" }
 ];
 
-const container = document.getElementById("productos-container");
-
-productos.forEach(producto => {
-  const col = document.createElement("div");
-  col.className = "col-12 col-sm-6 col-md-4";
-
-  col.innerHTML = `
-    <div class="card h-100">
-      <img src="${producto.imagen}" class="card-img-top product-img" alt="${producto.nombre}">
-      <div class="card-body">
-        <h5 class="card-title">${producto.nombre}</h5>
-        <p class="card-text">$${producto.precio}</p>
-        <button type="button" class="btn btn-primary btn-compra">Agregar al carrito</button>
-      </div>
-    </div>
-  `;
-
-  // Ahora buscamos el botón dentro del div 'col' que acabamos de crear
-  const botonAgregar = col.querySelector('button.btn-compra');
-
-  // Agregamos el event listener al botón para agregar el producto correspondiente
-  botonAgregar.addEventListener('click', () => {
-    agregarAlCarrito(producto);
-  });
-
-  // Finalmente agregamos el 'col' al container del DOM
-  container.appendChild(col);
-});
+mostrar_productos(productos);
